Topic controller business rules

+++ asp-net/API_Learn/API_Learn/Controllers/TopicController.cs
@@ -2,9 +2,12 @@
 using ApiCatalogo.Repositories;
 using DSLearn.Controllers.Utils;
 using DSLearn.Dtos;
+using DSLearn.Entities;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
+using System.Security.Claims;
 
 namespace DSLearn.Controllers
 {
@@ -14,10 +17,12 @@ namespace DSLearn.Controllers
     {
 
         private readonly IUnitOfWork _unitOfWork;
+        private readonly UserManager<User> _userManager;
 
-        public TopicController(IUnitOfWork unitOfWork)
+        public TopicController(IUnitOfWork unitOfWork, UserManager<User> userManager)
         {
             _unitOfWork = unitOfWork;
+            _userManager = userManager;
         }
 
 
@@ -41,9 +46,39 @@ namespace DSLearn.Controllers
 
 
         [HttpPost]
-        //[Authorize(Policy = "AdminOnly")]
+        [Authorize(Policy = "StudentOnly")]
         public async Task<ActionResult<dynamic>> Insert([FromBody] TopicInsertDTO dto)
         {
+
+            var userId = await _userManager.FindByIdAsync(dto.AuthorId);
+
+            if (userId == null)
+            {
+                return BadRequest($"Id {dto.AuthorId} does not exists");
+            }
+
+
+
+            // Obter as reivindicações de identidade do usuário a partir do contexto HTTP
+            var claimsIdentity = (ClaimsIdentity)User.Identity;
+
+            // Obter o ID único do usuário a partir das reivindicações de identidade
+            var userLoggedId = claimsIdentity.FindFirst(ClaimTypes.NameIdentifier).Value;
+
+            // Buscar informações do usuário utilizando o UserManager
+            var userLogged = await _userManager.FindByIdAsync(userLoggedId);
+
+
+            // Obter as roles (funções) associadas ao usuário utilizando o UserManager
+            var rolesUserLogged = await _userManager.GetRolesAsync(userLogged);
+
+            // Verificar se o ID do usuário não corresponde ao ID fornecido OU se o usuário não tem a função "Instructor"
+            if (!userLoggedId.Equals(dto.AuthorId, StringComparison.OrdinalIgnoreCase) && !rolesUserLogged.Contains("Instructor"))
+            {
+                // Se a condição acima for verdadeira, retornar uma resposta de não autorizado
+                return Unauthorized("Access Denied: The user does not have permission to make this task. Either the user ID does not match or the user does not have the 'Admin' role.");
+            }
+
             TopicDTO result = _unitOfWork.TopicRepository.Insert(dto);
             await _unitOfWork.CommitAsync();
 
@@ -51,11 +86,40 @@ namespace DSLearn.Controllers
         }
 
         [HttpPut("{id}")]
-       //[Authorize(Policy = "AdminOnly")]
+       //[Authorize(Policy = "StudentOnly")]
         public async Task<ActionResult<dynamic>> Update([FromBody] TopicInsertDTO dto, int id)
         {
             try
             {
+                var userId = await _userManager.FindByIdAsync(dto.AuthorId);
+
+                if (userId == null)
+                {
+                    return BadRequest($"Id {dto.AuthorId} does not exists");
+                }
+
+
+
+                // Obter as reivindicações de identidade do usuário a partir do contexto HTTP
+                var claimsIdentity = (ClaimsIdentity)User.Identity;
+
+                // Obter o ID único do usuário a partir das reivindicações de identidade
+                var userLoggedId = claimsIdentity.FindFirst(ClaimTypes.NameIdentifier).Value;
+
+                // Buscar informações do usuário utilizando o UserManager
+                var userLogged = await _userManager.FindByIdAsync(userLoggedId);
+
+
+                // Obter as roles (funções) associadas ao usuário utilizando o UserManager
+                var rolesUserLogged = await _userManager.GetRolesAsync(userLogged);
+
+                // Verificar se o ID do usuário não corresponde ao ID fornecido OU se o usuário não tem a função "Instructor"
+                if (!userLoggedId.Equals(dto.AuthorId, StringComparison.OrdinalIgnoreCase) && !rolesUserLogged.Contains("Instructor"))
+                {
+                    // Se a condição acima for verdadeira, retornar uma resposta de não autorizado
+                    return Unauthorized("Access Denied: The user does not have permission to make this task. Either the user ID does not match or the user does not have the 'Admin' role.");
+                }
+
                 TopicDTO result = _unitOfWork.TopicRepository.Update(dto, id);
                 await _unitOfWork.CommitAsync();
                 return Ok(result);
@@ -63,12 +127,41 @@ namespace DSLearn.Controllers
             catch (ArgumentException ex) { return ErrorMessages.ErrorMessage(id); }
         }
 
-        [HttpDelete("{id}")]
+        [HttpDelete("{queryUserId}/{id}")]
         //[Authorize(Policy = "AdminOnly")]
-        public async Task<ActionResult<dynamic>> Delete(int id)
+        public async Task<ActionResult<dynamic>> Delete(string queryUserId,int id)
         {
             try
             {
+                var userId = await _userManager.FindByIdAsync(queryUserId);
+
+                if (userId == null)
+                {
+                    return BadRequest($"Id {queryUserId} does not exists");
+                }
+
+
+
+                // Obter as reivindicações de identidade do usuário a partir do contexto HTTP
+                var claimsIdentity = (ClaimsIdentity)User.Identity;
+
+                // Obter o ID único do usuário a partir das reivindicações de identidade
+                var userLoggedId = claimsIdentity.FindFirst(ClaimTypes.NameIdentifier).Value;
+
+                // Buscar informações do usuário utilizando o UserManager
+                var userLogged = await _userManager.FindByIdAsync(userLoggedId);
+
+
+                // Obter as roles (funções) associadas ao usuário utilizando o UserManager
+                var rolesUserLogged = await _userManager.GetRolesAsync(userLogged);
+
+                // Verificar se o ID do usuário não corresponde ao ID fornecido OU se o usuário não tem a função "Instructor"
+                if (!userLoggedId.Equals(queryUserId, StringComparison.OrdinalIgnoreCase) && !rolesUserLogged.Contains("Instructor"))
+                {
+                    // Se a condição acima for verdadeira, retornar uma resposta de não autorizado
+                    return Unauthorized("Access Denied: The user does not have permission to make this task. Either the user ID does not match or the user does not have the 'Admin' role.");
+                }
+
                 _unitOfWork.TopicRepository.Delete(id);
                 await _unitOfWork.CommitAsync();
                 return NoContent();
